src/app/components/listado/listado.component.ts
import { CommonModule } from '@angular/common';
import { Component, OnInit, ViewChild } from '@angular/core';
import { ReactiveFormsModule } from '@angular/forms';
import { UserService } from '../../services/user.service';
import { AngularMaterialModule } from '../../modules/angular-material.module';
import { MatTableDataSource } from '@angular/material/table';
import { MatPaginator } from '@angular/material/paginator';
import { MatSort } from '@angular/material/sort';
import Swal from 'sweetalert2';
import { AuthService } from '../../services/auth.service';
import { Router } from '@angular/router';

@Component({
  selector: 'app-listado',
  imports: [CommonModule, ReactiveFormsModule, AngularMaterialModule],
  templateUrl: './listado.component.html',
  styleUrl: './listado.component.scss'
})
export class ListadoComponent implements OnInit {
  displayedColumns: string[] = ['name', 'email', 'phone'];
  dataSource: MatTableDataSource<any> = new MatTableDataSource();
  @ViewChild(MatPaginator) paginator!: MatPaginator;
  @ViewChild(MatSort) sort!: MatSort;

  constructor(
    private _userService: UserService,
    private _authService: AuthService,
    private _router: Router
  ) {}

  ngOnInit(): void {
    this.loadUsers();
  }

  loadUsers(): void {
    this._userService.listarUsuarios().subscribe({
      next: (users) => {
        this.dataSource.data = users;
        this.dataSource.paginator = this.paginator;
        this.dataSource.sort = this.sort;
      },
      error: (err) => {
        Swal.fire({
          icon: 'error',
          title: 'Error al cargar usuarios',
          text: 'Hubo un error interno del servidor. Intenta nuevamente.',
        });
      }
    });
  }

  cerrarSesion() {
    this._authService.clearToken();
    this._router.navigate(['/login']);
  }
}
